BackendIE/ml/requirements.txt
requests
PyPDF2
sentence-transformers
scikit-learn
xgboost
pandas
numpy
joblib
pdfminer.six
